reorganizing for packagist
