+++ app/DIContainerAdapter.php
@@ -3,7 +3,7 @@
 namespace LearningApp;
 
 use DI\Container;
-use Learning\IRouterContainer;
+use Tpab\Router\IRouterContainer;
 
 class DIContainerAdapter implements IRouterContainer
 {
@@ -26,7 +26,7 @@ class DIContainerAdapter implements IRouterContainer
     public function get($route, $parameters)
     {
         $to_return = $this->container->get($route);
-        
+
         if (is_callable($to_return)) {
             if (is_array($to_return)) {
                 $to_return[0] = $this->container->make($to_return[0], $parameters);
