+++ composer.json
@@ -1,24 +1,28 @@
 {
-    "name": "tpabarbosa/learning-router",
+    "name": "tpab/router",
     "description": "Project to learn how routers work",
-    "type": "project",
-    "require-dev": {
-        "phpunit/phpunit": "^8.5",
-        "mockery/mockery": "^1.3"
-    },
+    "type": "library",
     "license": "MIT",
     "authors": [{
         "name": "Tatiana Barbosa",
         "email": "tpabarbosa@yahoo.com.br"
     }],
     "require": {
+        "php": ">=7.0",
         "php-di/php-di": "^6.3"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^8.5"
     },
     "autoload": {
         "psr-4": {
-            "Learning\\": "src/",
-            "LearningTests\\": "tests/",
+            "Tpab\\Router\\": "src/",
             "LearningApp\\": "app/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Tpab\\Router\\Tests\\": "tests/"
         }
     }
 }
